+++ apps/marketplace/components/Brief/BriefResponseForm.js
@@ -45,110 +45,137 @@ const BriefResponseForm = ({
               submitClicked={submitClicked}
               setFocus={setFocus}
             />}
-
-          <div>
-            <div className={styles.stepTitle}>Specialist {specialistNumber} of 3</div>
-            <h1 className="uikit-display-6">{specialistName}</h1>
-            <div className="uikit-display-5">
-              About
-            </div>
-            <Form model={model} id="briefResponse" onSubmit={data => handleSubmit(data)}>
-              <Textfield
-                model={`${model}.availability`}
-                name="availability"
-                id="availability"
-                htmlFor="availability"
-                label="Earliest start date"
-                maxLength={100}
-                description={
-                  <span>
-                    Preferred start date is <b>{brief.startDate}</b>
-                  </span>
-                }
-                validators={{
-                  required
-                }}
-                messages={{
-                  required: 'Enter a date for when you can start the project'
-                }}
-              />
-              {brief.lotSlug &&
-                brief.lotSlug === 'digital-professionals' &&
+          {!specialistName ?
+            <div>
+              {!addAnotherSpecialist &&
+                <div>
+                  <h1 className="uikit-display-5">Add up to 3 specialists</h1>
+                  <div>You can withdraw or replace specialists any time before the opportunity closes on {format(new Date(brief.applicationsClosedAt), 'DD/MM/YYYY')}.</div>
+                  <br/>
+                </div>
+              }
+              <div className="uikit-display-4"><strong>Specialist {specialistNumber}</strong></div>
+              <Form model={model} id="briefName" onSubmit={data => handleNameSubmit(data.specialistName)}>
                 <Textfield
-                  model={`${model}.dayRate`}
-                  name="dayRate"
-                  id="dayRate"
-                  htmlFor="dayRate"
-                  label="Day rate (including GST)"
+                  model={`${model}.specialistName`}
+                  name="specialistName"
+                  id="specialistName"
+                  htmlFor="specialistName"
+                  label="Full name"
                   validators={{
-                    required,
-                    validPercentage
+                    required
                   }}
                   messages={{
-                    required: 'A day rate is required',
-                    validPercentage: 'Enter only numbers eg. 600.00'
+                    required: 'A name is required'
                   }}
-                />}
-              {app.supplierCode
-                ? brief.lotSlug &&
-                  brief.lotSlug === 'digital-professionals' &&
-                  <FilesInput
-                    label="Resume"
-                    hint="Attachments must be PDF or ODT format and a maximum of 5MB"
-                    name="attachedDocumentURL"
-                    model={model}
-                    formFields={1}
-                    fieldLabel="Upload resume"
-                    url={`/brief/${brief.id}/respond/documents/${app.supplierCode}`}
-                    api={dmapi}
-                    description=""
-                  />
-                : <PageAlert as="warning" setFocus={setFocus}>
-                    <h4>There was a problem loading your details</h4>
-                    <p>Only logged in sellers can respond to briefs</p>
-                  </PageAlert>}
-              <fieldset className={styles.x_uikit_fieldset}>
-                <h2>Skills and experience?</h2>
+                />
+                <input className="uikit-btn" type="submit" value="Start application" />
+              </Form>
+            </div>
+          :
+            <div>
+              <div className={styles.stepTitle}>Specialist {specialistNumber} of 3</div>
+              <h1 className="uikit-display-6">{specialistName}</h1>
+              <div className="uikit-display-5">
+                About
+              </div>
+              <Form model={model} id="briefResponse" onSubmit={data => handleSubmit(data)}>
+                <Textfield
+                  model={`${model}.availability`}
+                  name="availability"
+                  id="availability"
+                  htmlFor="availability"
+                  label="Earliest start date"
+                  maxLength={100}
+                  description={
+                    <span>
+                      Preferred start date is <b>{brief.startDate}</b>
+                    </span>
+                  }
+                  validators={{
+                    required
+                  }}
+                  messages={{
+                    required: 'Enter a date for when you can start the project'
+                  }}
+                />
                 {brief.lotSlug &&
                   brief.lotSlug === 'digital-professionals' &&
-                  <p>Answer the following criteria for all candidates</p>}
-                {brief.essentialRequirements &&
-                  brief.essentialRequirements.map((requirement, i) =>
-                    <Textarea
-                      key={`essentialRequirement.${i}`}
-                      model={`${model}.essentialRequirements[${i}]`}
-                      name={`essentialRequirement.${i}`}
-                      id={`essentialRequirement.${i}`}
-                      controlProps={{ limit: 150 }}
-                      label={requirement}
-                      validators={{ required }}
-                      showMessagesDuringFocus
-                      messages={{
-                        required: `This is an essential requirement, let the buyer know how the skills and experience criteria is met`
-                      }}
+                  <Textfield
+                    model={`${model}.dayRate`}
+                    name="dayRate"
+                    id="dayRate"
+                    htmlFor="dayRate"
+                    label="Day rate (including GST)"
+                    validators={{
+                      required,
+                      validPercentage
+                    }}
+                    messages={{
+                      required: 'A day rate is required',
+                      validPercentage: 'Enter only numbers eg. 600.00'
+                    }}
+                  />}
+                {app.supplierCode
+                  ? brief.lotSlug &&
+                    brief.lotSlug === 'digital-professionals' &&
+                    <FilesInput
+                      label="Resume"
+                      hint="Attachments must be PDF or ODT format and a maximum of 5MB"
+                      name="attachedDocumentURL"
+                      model={model}
+                      formFields={1}
+                      fieldLabel="Upload resume"
+                      url={`/brief/${brief.id}/respond/documents/${app.supplierCode}`}
+                      api={dmapi}
+                      description=""
                     />
-                  )}
-                {brief.niceToHaveRequirements &&
-                  brief.niceToHaveRequirements.map((requirement, i) =>
-                    <Textarea
-                      key={`niceToHaveRequirement.${i}`}
-                      model={`${model}.niceToHaveRequirements[${i}]`}
-                      name={`niceToHaveRequirement.${i}`}
-                      id={`niceToHaveRequirement.${i}`}
-                      controlProps={{ limit: 150 }}
-                      label={`${requirement} (optional)`}
-                    />
-                  )}
-              </fieldset>
-              {currentlySending
-                ? <LoadingButton />
-              : <input className="uikit-btn right-button-margin" type="submit" value="Submit specialist" onClick={submitClicked} />}
-              {currentlySending
-                ? <LoadingButton />
-              : <div className="uikit-btn uikit-btn--tertiary" onClick={addAnotherClicked(true)}>Submit and add another</div>}
-            </Form>
-          </div>
-
+                  : <PageAlert as="warning" setFocus={setFocus}>
+                      <h4>There was a problem loading your details</h4>
+                      <p>Only logged in sellers can respond to briefs</p>
+                    </PageAlert>}
+                <fieldset className={styles.x_uikit_fieldset}>
+                  <h2>Skills and experience?</h2>
+                  {brief.lotSlug &&
+                    brief.lotSlug === 'digital-professionals' &&
+                    <p>Answer the following criteria for all candidates</p>}
+                  {brief.essentialRequirements &&
+                    brief.essentialRequirements.map((requirement, i) =>
+                      <Textarea
+                        key={`essentialRequirement.${i}`}
+                        model={`${model}.essentialRequirements[${i}]`}
+                        name={`essentialRequirement.${i}`}
+                        id={`essentialRequirement.${i}`}
+                        controlProps={{ limit: 150 }}
+                        label={requirement}
+                        validators={{ required }}
+                        showMessagesDuringFocus
+                        messages={{
+                          required: `This is an essential requirement, let the buyer know how the skills and experience criteria is met`
+                        }}
+                      />
+                    )}
+                  {brief.niceToHaveRequirements &&
+                    brief.niceToHaveRequirements.map((requirement, i) =>
+                      <Textarea
+                        key={`niceToHaveRequirement.${i}`}
+                        model={`${model}.niceToHaveRequirements[${i}]`}
+                        name={`niceToHaveRequirement.${i}`}
+                        id={`niceToHaveRequirement.${i}`}
+                        controlProps={{ limit: 150 }}
+                        label={`${requirement} (optional)`}
+                      />
+                    )}
+                </fieldset>
+                {currentlySending
+                  ? <LoadingButton />
+                : <input className="uikit-btn right-button-margin" type="submit" value="Submit specialist" onClick={() => {submitClicked}} />}
+                {currentlySending
+                  ? <LoadingButton />
+                : <input className="uikit-btn uikit-btn--tertiary" type="button" value="Submit and add another" onClick={() => { addAnotherClicked(true) }} />}
+              </Form>
+            </div>
+          }
         </article>
       </div>
     </DocumentTitle>
